Move mixins to directory
