+++ src/lib/mixins/SemanticUIVueMixin/getChildProps.js
