+++ src/lib/mixins/SemanticUIVueMixin/getElementType.js
