+++ src/lib/mixins/index.js
@@ -1,0 +1,2 @@
+export * from './SemanticUIVueMixin';
+export * from './classMixin';
